Update: Reading history

+++ src/main/java/com/iuh/fit/readhub/dto/request/ReadingHistoryRequest.java
@@ -7,5 +7,5 @@ public class ReadingHistoryRequest {
     private Long readingHistoryId;
     private Long userId;
     private Long bookId;
-    private Integer timeSpent;
+    private Long timeSpent;
 }
